% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_markdowns.R
\name{render_markdowns}
\alias{render_markdowns}
\title{the render_markdowns function}
\usage{
render_markdowns(ids = projectids)
}
\arguments{
\item{ids}{defaults to projectids as defined in run_reports function}
}
\description{
This function allows you to render the parameterized reports
}
\examples{
render_markdowns()
}
\keyword{render}
